+++ src/App.js
@@ -2,34 +2,16 @@ import './css/reset.css';
 import './css/App.css';
 import CardList from './components/CardList';
 import { useCallback, useEffect, useState } from 'react';
-import { getCards, getUserProfile } from './api/api';
+import { getCards } from './api/api';
 import useAsync from './hooks/useAsync';
-import UserProfile from './components/UserProfile';
 import Folder from './components/Folder';
 import Footer from './components/Footer/Footer';
+import Nav from './components/Nav/Nav';
 
 function App() {
-  const [userProfile, setUserProfile] = useState(null);
   const [folderProfile, setFolderProfile] = useState(null);
   const [cards, setCards] = useState([]);
-  const [isLoadingUserProfile, userProfileLoadingError, getUserProfileAsync] = useAsync(getUserProfile);
   const [isLoadingCards, cardsLoadingError, getCardsAsync] = useAsync(getCards);
-
-  const handleUserProfile = useCallback(
-    async () => {
-      const result = await getUserProfileAsync();
-      if (!result) {
-        return;
-      }
-
-      const userProfile = { ...result };
-      setUserProfile(userProfile);
-    }, [getUserProfileAsync],
-  );
-
-  useEffect(() => {
-    handleUserProfile();
-  }, [handleUserProfile]);
 
   const handleLoad = useCallback(
     async () => {
@@ -43,7 +25,7 @@ function App() {
       setFolderProfile({
         avatar: owner?.profileImageSource ?? '',
         ownerName: owner?.name ?? '',
-        folderName: name
+        folderName: name,
       });
       setCards(cards);
     }, [getCardsAsync],
@@ -55,13 +37,7 @@ function App() {
 
   return (
     <div className='App'>
-      <nav>
-        <div className='gnb'>
-          <img className='logo' src='/assets/images/logo.svg' alt='홈으로 연결된 Linkbrary 로고' />
-          <UserProfile userProfile={userProfile} />
-          {userProfileLoadingError?.message && <span>{userProfileLoadingError.message}</span>}
-        </div>
-      </nav>
+      <Nav />
       <header>
         <Folder folderProfile={folderProfile} />
       </header>
